Bugfix for ANW-321: Conditions Governing Access and Conditions Governing Use are now mapped to <dc:rights>.
